Launched project list moved and bugs fixed
