+++ app/common/src/services/Backend.ts
@@ -295,7 +295,7 @@ export interface BaseProject {
 export interface CreatedProject extends BaseProject {
   readonly state: ProjectStateType
   readonly packageName: string
-  readonly ensoPath?: EnsoPath
+  readonly ensoPath: EnsoPath
 }
 
 /** A `Project` returned by `updateProject`. */
