+++ app/gui/src/dashboard/pages/dashboard/Dashboard/index.ts
@@ -1,3 +1,2 @@
 import Dashboard from './Dashboard'
-export { dataLoader } from './dataLoader'
 export default Dashboard
