--- app/gui/src/dashboard/providers/ProjectsProvider.tsx
@@ -1,204 +1,0 @@
-/** @file The React provider (and associated hooks) for providing reactive events. */
-import * as React from 'react'
-
-import invariant from 'tiny-invariant'
-import * as z from 'zod'
-
-import * as eventCallbacks from '#/hooks/eventCallbackHooks'
-import { useLocalStorageState } from '#/hooks/localStoreState'
-import * as backendModule from '#/services/Backend'
-import * as array from '#/utilities/array'
-import LocalStorage from '#/utilities/LocalStorage'
-import { reactive } from 'vue'
-
-const TAB_TYPES = ['drive', 'settings'] as const
-
-/** Main content of the screen. Only one should be visible at a time. */
-export type TabType = (typeof TAB_TYPES)[number]
-
-declare module '#/utilities/LocalStorage' {
-  /** */
-  interface LocalStorageData {
-    readonly isAssetPanelVisible: boolean
-    readonly page: z.infer<typeof PAGES_SCHEMA>
-    readonly launchedProjects: z.infer<typeof LAUNCHED_PROJECT_SCHEMA>
-  }
-}
-
-const PROJECT_ID_SCHEMA = z.custom<backendModule.ProjectId>(
-  (x) => typeof x === 'string' && x.startsWith('project-'),
-)
-const DIRECTORY_ID_SCHEMA = z.custom<backendModule.DirectoryId>(
-  (x) => typeof x === 'string' && x.startsWith('directory-'),
-)
-const PROJECT_SCHEMA = z
-  .object({
-    id: PROJECT_ID_SCHEMA,
-    parentId: DIRECTORY_ID_SCHEMA,
-    title: z.string(),
-    ensoPath: z.string(),
-    type: z.nativeEnum(backendModule.BackendType),
-    preventAutoReopen: z.boolean().optional(),
-    hybrid: z.optional(
-      z.object({
-        cloudProjectId: PROJECT_ID_SCHEMA,
-        cloudParentId: DIRECTORY_ID_SCHEMA,
-        parentId: DIRECTORY_ID_SCHEMA,
-        cloudProjectDirectoryPath: z.string(),
-      }),
-    ),
-  })
-  .readonly()
-const LAUNCHED_PROJECT_SCHEMA = z.array(PROJECT_SCHEMA).readonly()
-
-/** Launched project information. */
-export type LaunchedProject = z.infer<typeof PROJECT_SCHEMA>
-/** Launched project ID. */
-export type LaunchedProjectId = backendModule.ProjectId
-
-LocalStorage.registerKey('launchedProjects', {
-  isUserSpecific: true,
-  schema: LAUNCHED_PROJECT_SCHEMA,
-})
-
-// eslint-disable-next-line react-refresh/only-export-components
-export const PAGES_SCHEMA = z
-  .enum(TAB_TYPES)
-  .or(
-    z.custom<LaunchedProjectId>(
-      (value) => typeof value === 'string' && value.startsWith('project-'),
-    ),
-  )
-
-LocalStorage.registerKey('page', { schema: PAGES_SCHEMA })
-
-/** State contained in a `ProjectsContext`. */
-export interface ProjectsContextType {
-  readonly setLaunchedProjects: (launchedProjects: readonly LaunchedProject[]) => void
-  readonly addLaunchedProject: (project: LaunchedProject) => void
-  readonly removeLaunchedProject: (projectId: LaunchedProjectId) => void
-  readonly updateLaunchedProjects: (
-    update: (projects: readonly LaunchedProject[]) => readonly LaunchedProject[],
-  ) => void
-  readonly getState: () => {
-    readonly launchedProjects: readonly LaunchedProject[]
-  }
-}
-
-const ProjectsContext = React.createContext<ProjectsContextType | null>(null)
-const LaunchedProjectsContext = React.createContext<readonly LaunchedProject[] | null>(null)
-
-/** Props for a {@link ProjectsProvider}. */
-export type ProjectsProviderProps = Readonly<React.PropsWithChildren>
-
-/**
- * A React provider (and associated hooks) for determining whether the current area
- * containing the current element is focused.
- */
-export default function ProjectsProvider(props: ProjectsProviderProps) {
-  const { children } = props
-
-  const [launchedProjects, setLaunchedProjects] = useLocalStorageState(
-    'launchedProjects',
-    reactive(array.EMPTY_ARRAY),
-    {
-      sanitize: (savedLaunchedProjects) =>
-        savedLaunchedProjects.map((project) => {
-          if (project.type === backendModule.BackendType.local && project.hybrid == null) {
-            return project
-          } else {
-            // Disallow Cloud projects and Hybrid projects from auto-opening
-            return { ...project, preventAutoReopen: true }
-          }
-        }),
-    },
-  )
-
-  const addLaunchedProject = eventCallbacks.useEventCallback((project: LaunchedProject) => {
-    setLaunchedProjects((current) => [...current, project])
-  })
-  const removeLaunchedProject = eventCallbacks.useEventCallback((projectId: LaunchedProjectId) => {
-    setLaunchedProjects((current) =>
-      current.filter(({ id, hybrid }) => id !== projectId && hybrid?.cloudProjectId !== projectId),
-    )
-  })
-  const updateLaunchedProjects = eventCallbacks.useEventCallback(
-    (update: (projects: readonly LaunchedProject[]) => readonly LaunchedProject[]) => {
-      setLaunchedProjects((current) => update(current))
-    },
-  )
-
-  const getState = eventCallbacks.useEventCallback(() => ({
-    launchedProjects,
-  }))
-
-  const projectsContextValue = {
-    updateLaunchedProjects,
-    addLaunchedProject,
-    removeLaunchedProject,
-    setLaunchedProjects,
-    getState,
-  }
-
-  return (
-    <ProjectsContext.Provider value={projectsContextValue}>
-      <LaunchedProjectsContext.Provider value={launchedProjects}>
-        {children}
-      </LaunchedProjectsContext.Provider>
-    </ProjectsContext.Provider>
-  )
-}
-
-/** The projects store. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useProjectsStore() {
-  const context = React.useContext(ProjectsContext)
-
-  invariant(context != null, 'Projects store can only be used inside an `ProjectsProvider`.')
-
-  return context
-}
-
-/** Returns the launched projects context. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useLaunchedProjects() {
-  const context = React.useContext(LaunchedProjectsContext)
-
-  invariant(
-    context != null,
-    'Launched projects context can only be used inside an `ProjectsProvider`.',
-  )
-
-  return context
-}
-
-/** A function to update launched projects. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useUpdateLaunchedProjects() {
-  const { updateLaunchedProjects } = useProjectsStore()
-  return updateLaunchedProjects
-}
-
-/** A function to add a new launched project. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useAddLaunchedProject() {
-  const { addLaunchedProject } = useProjectsStore()
-  return addLaunchedProject
-}
-
-/** A function to remove a launched project. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useRemoveLaunchedProject() {
-  const { removeLaunchedProject } = useProjectsStore()
-  return removeLaunchedProject
-}
-
-/** A function to remove all launched projects. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useClearLaunchedProjects() {
-  const { setLaunchedProjects } = useProjectsStore()
-
-  return eventCallbacks.useEventCallback(() => {
-    setLaunchedProjects([])
-  })
-}
