--- app/gui/src/providers/react/rightPanel.tsx
@@ -1,37 +1,0 @@
-import { RightPanelData, useRightPanelData as useRightPanelDataVue } from '$/providers/rightPanel'
-import * as react from 'react'
-import { applyPureReactInVue } from 'veaury'
-import { useInReactFunction, useVueValue } from './common'
-
-const RightPanelDataContext = react.createContext<RightPanelData | null>(null)
-export const useRightPanelData = useInReactFunction(RightPanelDataContext)
-
-export const RightPanelDataProviderForReact = applyPureReactInVue(
-  ({ value, children }: react.PropsWithChildren<{ value: RightPanelData }>) => {
-    return <RightPanelDataContext.Provider value={value}>{children}</RightPanelDataContext.Provider>
-  },
-  {
-    useInjectPropsFromWrapper: () => {
-      const result = {
-        value: useRightPanelDataVue(),
-      }
-      // Avoid annoying warning about __veauryInjectedProps__ property by returning a function.
-      return () => result
-    },
-  },
-)
-
-/**
- * A hook to read currently focused asset for right panel, e.g. the currently selected asset
- * in Drive View.
- */
-export function useRightPanelFocusedAsset() {
-  const rightPanel = useRightPanelData()
-  return useVueValue(react.useCallback(() => rightPanel.focusedAsset, [rightPanel]))
-}
-
-/** A hook reading current category set for right panel context. */
-export function useRightPanelContextCategory() {
-  const rightPanel = useRightPanelData()
-  return useVueValue(react.useCallback(() => rightPanel.context?.category, [rightPanel]))
-}
